initial libretro port.

+++ source/getset.h
@@ -120,7 +120,7 @@ uint8 S9xGetByte (uint32 Address)
 		return (*(GetAddress + (Address & 0xffff)));
     }
 	
-    switch ((int) GetAddress)
+    switch ((intptr_t) GetAddress)
     {
     case CMemory::MAP_PPU:
 		return (S9xGetPPU (Address & 0xffff));
@@ -224,7 +224,7 @@ uint16 S9xGetWord (uint32 Address)
 #endif	
     }
 
-    switch ((int) GetAddress)
+    switch ((intptr_t) GetAddress)
     {
     case CMemory::MAP_PPU:
 		return (S9xGetPPU (Address & 0xffff) |
@@ -347,7 +347,7 @@ void S9xSetByte (uint8 Byte, uint32 Address)
 		return;
     }
 	
-    switch ((int) SetAddress)
+    switch ((intptr_t) SetAddress)
     {
     case CMemory::MAP_PPU:
 		S9xSetPPU (Byte, Address & 0xffff);
@@ -480,7 +480,7 @@ void S9xSetWord (uint16 Word, uint32 Address)
 		return;
     }
 	
-    switch ((int) SetAddress)
+    switch ((intptr_t) SetAddress)
     {
     case CMemory::MAP_PPU:
 		S9xSetPPU ((uint8) Word, Address & 0xffff);
@@ -601,7 +601,7 @@ uint8 *GetBasePointer (uint32 Address)
 	{
 		return s7r.bank50;
 	}
-    switch ((int) GetAddress)
+    switch ((intptr_t) GetAddress)
     {
 	case CMemory::MAP_SPC7110_DRAM:
 #ifdef SPC7110_DEBUG
@@ -667,7 +667,7 @@ uint8 *S9xGetMemPointer (uint32 Address)
 	if(Settings.SPC7110&&((Address&0x7FFFFF)==0x4800))
 		return s7r.bank50;
 
-    switch ((int) GetAddress)
+    switch ((intptr_t) GetAddress)
     {
 	case CMemory::MAP_SPC7110_DRAM:
 #ifdef SPC7110_DEBUG
@@ -727,7 +727,7 @@ void S9xSetPCBase (uint32 Address)
 		return;
     }
 	
-    switch ((int) GetAddress)
+    switch ((intptr_t) GetAddress)
     {
     case CMemory::MAP_PPU:
 		CPU.PCBase = Memory.FillRAM;
